+++ docs/known_issues.rst
@@ -21,3 +21,11 @@ Using equation 187 in `Calabretta & Greisen 2002 <https://www.aanda.org/articles
 
 Note that since these pixels are extremely rectangular, with an aspect ratio of ~3e-10, the cross terms in the PCij matrix are quite small: -3.4e-12 and -3.8e-7.
 Hopefully, 64-bit floats have enough precision to enable this to work all of the time.
+
+Raster gWCS Inverse and Time
+==============================
+
+The raster gWCS exposes five world coordinates: wavelength, helioprojective longitude, helioprojective latitude, elapsed time, and an explicit scan-step index.
+Because time may not be monotonic across flipped or repeated rasters, the WCS inverse (``world_to_array_index``) deliberately ignores the time component and resolves pixels using sky position plus the explicit scan-step coordinate only.
+Consequently, passing a world tuple with a mismatched time value to ``world_to_array_index`` or ``crop`` will silently return the same pixel as if the correct time had been supplied.
+To avoid confusion, prefer building crop bounds with ``cube.wcs.array_index_to_world`` rather than hand-constructing world tuples with arbitrary time values.
